Add confirmation to purge script

+++ scripts/purge.sh
@@ -4,9 +4,15 @@
 
 set -eu
 
+read -p "Are you sure you want to purge all docker networks, containers, images and volumes? (y/n) " answer
+if [ "$answer" != "y" ]; then
+    echo "Aborting..."
+    exit 1
+fi
+
 docker network prune -f
 docker container prune -f
-docker image prune -a -f
+docker image prune -af
 docker volume prune -f
 
 echo "Done."
